Revert "feature:migrate radio buttons to shadcn radio-group (#240)" (#254)
This reverts commit 517389c4f604ea4836e09fe5ea3832922835cfa4.

The version of `apps/next/components/ui/radio-group.tsx` added in #240
is not actually the Shadcn component. Having looked into it, it also
doesn't really make sense to add the Radix radio button package at all
right now, as it's only really suited to render a classic style of radio
button, which we currently don't use.

After converting the few remaining forms to Shadcn (or removing them, as
we're planning to do with onboarding and roles management), we can clean
up the `RadioButtons` component to render only the actual buttons rather
than the label and help text as well.

---------

Co-authored-by: Maya <19721695+code-elf@users.noreply.github.com>